lambda/requirements.txt
fastapi==0.110.0
mangum==0.17.0
pydantic==2.6.2
boto3==1.34.69
langchain-aws==0.1.2
langchain-core==0.1.28
langchain-community>=0.0.24
langchain-text-splitters==0.0.1
colorama==0.4.6
python-dotenv==1.0.1
faiss-cpu==1.7.4